+++ aws-java-sdk-devopsguru/src/main/java/com/amazonaws/services/devopsguru/AmazonDevOpsGuru.java
@@ -179,6 +179,32 @@ public interface AmazonDevOpsGuru {
 
     /**
      * <p>
+     * Returns the most recent feedback submitted in the current AWS account and Region.
+     * </p>
+     * 
+     * @param describeFeedbackRequest
+     * @return Result of the DescribeFeedback operation returned by the service.
+     * @throws AccessDeniedException
+     *         You don't have permissions to perform the requested operation. The user or role that is making the
+     *         request must have at least one IAM permissions policy attached that grants the required permissions. For
+     *         more information, see <a href="https://docs.aws.amazon.com/IAM/latest/UserGuide/access.html">Access
+     *         Management</a> in the <i>IAM User Guide</i>.
+     * @throws InternalServerException
+     *         An internal failure in an Amazon service occurred.
+     * @throws ResourceNotFoundException
+     *         A requested resource could not be found
+     * @throws ThrottlingException
+     *         The request was denied due to a request throttling.
+     * @throws ValidationException
+     *         Contains information about data passed in to a field during a request that is not valid.
+     * @sample AmazonDevOpsGuru.DescribeFeedback
+     * @see <a href="http://docs.aws.amazon.com/goto/WebAPI/devops-guru-2020-12-01/DescribeFeedback" target="_top">AWS
+     *      API Documentation</a>
+     */
+    DescribeFeedbackResult describeFeedback(DescribeFeedbackRequest describeFeedbackRequest);
+
+    /**
+     * <p>
      * Returns details about an insight that you specify using its ID.
      * </p>
      *